Extracted WebSocket registration
