+++ server.js
@@ -48,21 +48,22 @@ const utils = require('./lib/utils')
 const insecurity = require('./lib/insecurity')
 const models = require('./models')
 const datacreator = require('./data/datacreator')
-const notifications = require('./data/datacache').notifications
 const app = express()
 const server = require('http').Server(app)
-const io = require('socket.io')(server)
 const appConfiguration = require('./routes/appConfiguration')
 const captcha = require('./routes/captcha')
 const trackOrder = require('./routes/trackOrder')
 const config = require('config')
-let firstConnectedSocket = null
 
-global.io = io
 errorhandler.title = 'Juice Shop (Express ' + utils.version('express') + ')'
 
 require('./lib/validateConfig')()
 require('./lib/cleanupFtpFolder')()
+
+/* Locals */
+app.locals.captchaId = 0
+app.locals.captchaReqId = 1
+app.locals.captchaBypassReqTimes = []
 
 /* Bludgeon solution for possible CORS problems: Allow everything! */
 app.options('*', cors())
@@ -115,6 +116,10 @@ app.use(bodyParser.json())
 let accessLogStream = require('file-stream-rotator').getStream({filename: './access.log', frequency: 'daily', verbose: false, max_logs: '2d'})
 app.use(morgan('combined', {stream: accessLogStream}))
 
+/* Rate limiting */
+app.enable('trust proxy')
+app.use('/rest/user/reset-password', new RateLimit({ windowMs: 5 * 60 * 1000, max: 100, keyGenerator ({headers, ip}) { return headers['X-Forwarded-For'] || ip }, delayMs: 0 }))
+
 /** Authorization **/
 /* Checks on JWT in Authorization header */
 app.use(verify.jwtChallenges())
@@ -165,25 +170,11 @@ app.post('/api/Feedbacks', verify.captchaBypassChallenge())
 /* Unauthorized users are not allowed to access B2B API */
 app.use('/b2b/v2', insecurity.isAuthorized())
 
-/* Verifying DB related challenges can be postponed until the next request for challenges is coming via sequelize-restful */
+/* Verifying DB related challenges can be postponed until the next request for challenges is coming via epilogue */
 app.use(verify.databaseRelatedChallenges())
 
-const endpointLimiter = new RateLimit({
-  windowMs: 5 * 60 * 1000, /* 100 requests per 5 minutes */
-  max: 100,
-  keyGenerator ({headers, ip}) {
-    return headers['X-Forwarded-For'] || ip
-  },
-  delayMs: 0
-})
-
-app.enable('trust proxy')
-app.use('/rest/user/reset-password', endpointLimiter)
-
-epilogue.initialize({
-  app,
-  sequelize: models.sequelize
-})
+/* Generated API endpoints */
+epilogue.initialize({ app, sequelize: models.sequelize })
 
 const autoModels = ['User', 'Product', 'Feedback', 'BasketItem', 'Challenge', 'Complaint', 'Recycle', 'SecurityQuestion', 'SecurityAnswer']
 
@@ -202,11 +193,6 @@ for (const modelName of autoModels) {
     return context.continue
   })
 }
-
-// routes for the NoSql parts of the application
-app.get('/rest/product/:id/reviews', showProductReviews())
-app.put('/rest/product/:id/reviews', createProductReviews())
-app.patch('/rest/product/reviews', insecurity.isAuthorized(), updateProductReviews())
 
 /* Custom Restful API */
 app.post('/rest/user/login', login())
@@ -228,23 +214,26 @@ app.get('/rest/admin/application-version', appVersion())
 app.get('/redirect', redirect())
 app.get('/rest/captcha', captcha())
 app.post('/rest/track-order', trackOrder())
+
+/* NoSQL API endpoints */
+app.get('/rest/product/:id/reviews', showProductReviews())
+app.put('/rest/product/:id/reviews', createProductReviews())
+app.patch('/rest/product/reviews', insecurity.isAuthorized(), updateProductReviews())
+
 /* B2B Order API */
 app.post('/b2b/v2/orders', b2bOrder())
 
 /* File Upload */
 app.post('/file-upload', upload.single('file'), fileUpload())
+
 /* File Serving */
 app.get('/the/devs/are/so/funny/they/hid/an/easter/egg/within/the/easter/egg', easterEgg())
 app.get('/this/page/is/hidden/behind/an/incredibly/high/paywall/that/could/only/be/unlocked/by/sending/1btc/to/us', premiumReward())
 app.use(angular())
+
 /* Error Handling */
 app.use(verify.errorHandlingChallenge())
 app.use(errorhandler())
-/* Captcha Id */
-app.locals.captchaId = 0
-/* Captcha Bypass Challenge parameters */
-app.locals.captchaReqId = 1
-app.locals.captchaBypassReqTimes = []
 
 exports.start = async function (readyCallback) {
   await models.sequelize.sync({ force: true })
@@ -252,7 +241,7 @@ exports.start = async function (readyCallback) {
 
   server.listen(process.env.PORT || config.get('server.port'), () => {
     console.log(colors.yellow('Server listening on port %d'), config.get('server.port'))
-    registerWebsocketEvents()
+    require('./lib/registerWebsocketEvents')(server)
     if (readyCallback) {
       readyCallback()
     }
@@ -260,26 +249,6 @@ exports.start = async function (readyCallback) {
 
   require('./lib/populateIndexTemplate')()
   require('./lib/populateThreeJsTemplate')()
-}
-
-function registerWebsocketEvents () {
-  io.on('connection', socket => {
-    if (firstConnectedSocket === null) {
-      socket.emit('server started')
-      firstConnectedSocket = socket.id
-    }
-
-    notifications.forEach(notification => {
-      socket.emit('challenge solved', notification)
-    })
-
-    socket.on('notification received', data => {
-      const i = notifications.findIndex(({flag}) => flag === data)
-      if (i > -1) {
-        notifications.splice(i, 1)
-      }
-    })
-  })
 }
 
 exports.close = function (exitCode) {
